feat: create auth layout
