+++ client/src/components/layout/authLayout/index.tsx
@@ -1,0 +1,20 @@
+import { FC, ReactNode } from "react";
+import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
+import { redirect } from "next/navigation";
+import { cookies } from "next/headers";
+
+type AuthLayoutProps = {
+  children: ReactNode;
+};
+
+export const AuthLayout: FC<AuthLayoutProps> = async ({ children }) => {
+  const supabase = createServerComponentClient({ cookies });
+  const {
+    data: { session },
+  } = await supabase.auth.getSession();
+  if (!session) {
+    redirect("/sign_in");
+  }
+
+  return <>{children}</>;
+};
